requirements.txt
icrawler==0.6.7
judini==0.0.8
openai==0.27.8
python-dotenv==1.0.0
python_pptx==0.6.21
streamlit==1.23.1
streamlit_chat==0.0.2.2
